Name colisions don't cause errors anymore
The new cache system does not allow rewriting the CacheTable entry
by name, so it is necessary to check if the entry exists and
change its content instead of assigning new value to it.

The cache related code has been moved to nesm/cache.nim to reduce
the size of nesm.nim file.