Skip typedescs in nonIntrusiveBody
As described in https://github.com/nim-lang/Nim/pull/22535, there is a Nim bug where generic parameters used in templates act like values of types rather than types themselves. Fixing this causes this macro to fail on calls from `nonIntrusiveTemplate` because it now provides a typedesc for the first argument and a complaint is made that serializing typedescs is not possible. To avoid this, skip the typedesc representation.